requirements.txt
numpy
pandas
requests
datetime
yt-dlp
gspread
oauth2client
python-dotenv